Use axios-base for proxy
