+++ src/notifiers/apprise.ts
@@ -1,8 +1,6 @@
-import axios from 'axios';
-
+import axios from '../axios-base.js';
 import { NotifierService } from './notifier-service.js';
 import logger from '../common/logger.js';
-import { getNotifierAxiosConfig } from '../notify-axios-config.js';
 
 import type { AppriseConfig } from '../common/config/index.js';
 import type { NotificationReason } from '../interfaces/notification-reason.js';
@@ -37,13 +35,9 @@ url: ${url}`,
     L.trace({ apiUrl: this.config.apiUrl, jsonPayload }, 'Sending json payload');
 
     try {
-      await axios.post(
-        `${this.config.apiUrl}/notify`,
-        jsonPayload,
-        getNotifierAxiosConfig({
-          responseType: 'text',
-        }),
-      );
+      await axios.post(`${this.config.apiUrl}/notify`, jsonPayload, {
+        responseType: 'text',
+      });
     } catch (err) {
       L.error(err);
       L.error({ urls: this.config.urls }, `Failed to send message`);
