+++ src/notifiers/discord.ts
@@ -1,8 +1,6 @@
-import axios from 'axios';
-
+import axios from '../axios-base.js';
 import { NotifierService } from './notifier-service.js';
 import logger from '../common/logger.js';
-import { getNotifierAxiosConfig } from '../notify-axios-config.js';
 
 import type { DiscordConfig } from '../common/config/index.js';
 import type { NotificationReason } from '../interfaces/notification-reason.js';
@@ -49,9 +47,9 @@ export class DiscordNotifier extends NotifierService {
             },
           ],
         },
-        getNotifierAxiosConfig({
+        {
           responseType: 'json',
-        }),
+        },
       );
     } catch (err) {
       L.error(err);
